+++ hdi/dimensionality_reduction/gpgpu_sne/gpgpu_sne_compute.cpp
@@ -366,7 +366,7 @@ namespace hdi {
       glBindBufferBase(GL_SHADER_STORAGE_BUFFER, 5, _compute_buffers[GRADIENTS]);
 
       // Compute the gradients of the KL function
-      unsigned int grid_size = sqrt(num_points) + 1;
+      unsigned int grid_size = static_cast<unsigned int>(std::sqrt(num_points)) + 1;
       glDispatchCompute(grid_size, grid_size, 1);
 
       glMemoryBarrier(GL_SHADER_STORAGE_BARRIER_BIT);
@@ -396,7 +396,7 @@ namespace hdi {
 
       // Update the points
       unsigned int num_workgroups = (num_points * 2 / 64) + 1;
-      unsigned int grid_size = sqrt(num_workgroups) + 1;
+      unsigned int grid_size = static_cast<unsigned int>(std::sqrt(num_workgroups)) + 1;
       glDispatchCompute(grid_size, grid_size, 1);
     }
     template void GpgpuSneCompute<std::vector<hdi::data::SparseVec<uint32_t, float>>>::updatePoints(unsigned int, float*, embedding_type*, float, float);
@@ -424,7 +424,7 @@ namespace hdi {
 
       // Compute the bounds
       unsigned int num_workgroups = (num_points / 128) + 1;
-      unsigned int grid_size = sqrt(num_workgroups) + 1;
+      unsigned int grid_size = static_cast<unsigned int>(std::sqrt(num_workgroups)) + 1;
       glDispatchCompute(grid_size, grid_size, 1);
     }
     template void GpgpuSneCompute<std::vector<hdi::data::SparseVec<uint32_t, float>>>::updateEmbedding(unsigned int, float, float, float);
